+++ src/client.ts
@@ -34,7 +34,7 @@ import {
   Nodes,
   Status,
 } from './resources/nodes';
-import { VmLogsParams, VmLogsResponse, VmSSHParams, VmSSHResponse, Vms } from './resources/vms/vms';
+import { VMLogsParams, VMLogsResponse, VMSSHParams, VMSSHResponse, VMs } from './resources/vms/vms';
 import { type Fetch } from './internal/builtin-types';
 import { HeadersLike, NullableHeaders, buildHeaders } from './internal/headers';
 import { FinalRequestOptions, RequestOptions } from './internal/request-options';
@@ -88,7 +88,7 @@ export interface ClientOptions {
    * The maximum number of times that the client will retry a request in case of a
    * temporary failure, like a network error or a 5XX error from the server.
    *
-   * @default 2
+   * @default 0
    */
   maxRetries?: number | undefined;
 
@@ -149,7 +149,7 @@ export class SFCNodes {
    * @param {number} [opts.timeout=1 minute] - The maximum amount of time (in milliseconds) the client will wait for a response before timing out.
    * @param {MergedRequestInit} [opts.fetchOptions] - Additional `RequestInit` options to be passed to `fetch` calls.
    * @param {Fetch} [opts.fetch] - Specify a custom `fetch` function implementation.
-   * @param {number} [opts.maxRetries=2] - The maximum number of times the client will retry a request.
+   * @param {number} [opts.maxRetries=0] - The maximum number of times the client will retry a request.
    * @param {HeadersLike} opts.defaultHeaders - Default headers to include with every request to the API.
    * @param {Record<string, string | undefined>} opts.defaultQuery - Default query parameters to include with every request to the API.
    */
@@ -175,7 +175,7 @@ export class SFCNodes {
       parseLogLevel(readEnv('SFC_NODES_LOG'), "process.env['SFC_NODES_LOG']", this) ??
       defaultLogLevel;
     this.fetchOptions = options.fetchOptions;
-    this.maxRetries = options.maxRetries ?? 2;
+    this.maxRetries = options.maxRetries ?? 0;
     this.fetch = options.fetch ?? Shims.getDefaultFetch();
     this.#encoder = Opts.FallbackEncoder;
 
@@ -722,22 +722,22 @@ export class SFCNodes {
 
   static toFile = Uploads.toFile;
 
-  vms: API.Vms = new API.Vms(this);
+  vms: API.VMs = new API.VMs(this);
   nodes: API.Nodes = new API.Nodes(this);
 }
 
-SFCNodes.Vms = Vms;
+SFCNodes.VMs = VMs;
 SFCNodes.Nodes = Nodes;
 
 export declare namespace SFCNodes {
   export type RequestOptions = Opts.RequestOptions;
 
   export {
-    Vms as Vms,
-    type VmLogsResponse as VmLogsResponse,
-    type VmSSHResponse as VmSSHResponse,
-    type VmLogsParams as VmLogsParams,
-    type VmSSHParams as VmSSHParams,
+    VMs as VMs,
+    type VMLogsResponse as VMLogsResponse,
+    type VMSSHResponse as VMSSHResponse,
+    type VMLogsParams as VMLogsParams,
+    type VMSSHParams as VMSSHParams,
   };
 
   export {