+++ src/resources/vms/index.ts
@@ -1,10 +1,19 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 export {
+  Images,
+  type ImageCompleteUploadResponse,
+  type ImageGetResponse,
+  type ImageStartUploadResponse,
+  type ImageUploadResponse,
+  type ImageStartUploadParams,
+  type ImageUploadParams,
+} from './images';
+export {
   Script,
   type UserData,
   type ScriptCreateResponse,
   type ScriptRetrieveResponse,
   type ScriptCreateParams,
 } from './script';
-export { Vms, type VmLogsResponse, type VmSSHResponse, type VmLogsParams, type VmSSHParams } from './vms';
+export { VMs, type VMLogsResponse, type VMSSHResponse, type VMLogsParams, type VMSSHParams } from './vms';